Use float for current/expected values
Converting the float from json to decimal will make it consider its arithmetic
issues and the check for coverage would fail. For example, 95.65 in the json
would be converted to 95.650000000000005684341886080801486968994140625, which
is almost equal to the original value but greater in the comparison.

Also improve the fixes output to the value expected in the json

Change-Id: Iea1df470bafb2415a7cc80d45f5123572aa7b657