Default value for CR on gitlab/gerrit integration
If a Gerrit user comments in a patch and he doesn't
have the permission to make code reviews, the Gerrit
API doesn't return a code review value for his
participation. To fix that, we make a default value
return if the get for the CR dict doesn't found
the 'value' key.

Fixes: https://gitlab.com/stoqtech/private/bdil/issues/570
Change-Id: I0cc59d321d97da9edeb04d5cf84bbca22f79351d