Ignore flake8 rule E252
Some projects like `stoq-plugin-password` uses 2 different
PEP8 linters, one for pre-commit `flake8` and one for make coverage
`pyflakes3`.

The configurations of those 2 linters are not the same though.
So we can end up with a code that is ok on pre-commit but will break the
linter check later, on the test job.

One example of that is how default value for arguments in functions
with type hints works. flake8 expects spaces around the equal sign
while pyflakes3 dont. This is enforced by rule E252.

This patch includes an argument to ignore this rule on pre-commit.